LayoutTests/imported/w3c/web-platform-tests/workers/baseurl/alpha/importScripts-in-worker-expected.txt


FAIL Untitled assert_equals: expected "gamma/script.js" but got "beta/script.js"
